backend/requirements.txt
autoenv==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.1
urllib3==1.25.7
Werkzeug==0.16.0
